replaced all bigInts have been replaced with Numeric

+++ twitoff/models.py
@@ -6,7 +6,7 @@ DB = SQLAlchemy()
 
 
 class User( DB.Model):
-	id = DB.Column( DB.BigInteger, primary_key= True)
+	id = DB.Column( DB.Numeric( 20, 0), primary_key= True)
 	name = DB.Column( DB.String( 20), nullable= False)
 	newest_tweet_id = DB.Column( DB.Numeric( 20, 0))
 
@@ -18,7 +18,7 @@ class Tweet( DB.Model):
 	id = DB.Column( DB.Numeric( 20, 0), primary_key= True)
 	text = DB.Column( DB.Unicode( 500), nullable= False)
 	embedding = DB.Column( DB.PickleType, nullable= False)
-	user_id = DB.Column( DB.BigInteger, DB.ForeignKey( 'user.id'), nullable= False)
+	user_id = DB.Column( DB.Numeric( 20, 0), DB.ForeignKey( 'user.id'), nullable= False)
 	user = DB.relationship( "User", backref= DB.backref( 'tweets', lazy= True))
 
 	def __repr__( self):
